fix: add back container execution detection
